+++ src/mnelab/mainwindow.py
@@ -465,23 +465,6 @@ class MainWindow(QMainWindow):
             item.setFlags(item.flags() | Qt.ItemIsEditable)
             self.sidebar.setItem(row_index, 0, item)
 
-            # add delete button
-            delete_button = QPushButton(self)
-            icon_path = os.path.join(
-                os.path.dirname(__file__),
-                "icons",
-                "close-data.svg",
-            )
-
-            delete_button.setIcon(QIcon(icon_path))
-            delete_button.setStyleSheet(
-                "background: transparent; border: none; margin: auto;"
-            )
-            delete_button.clicked.connect(
-                lambda _, index=row_index: self.model.remove_data_at(index)
-            )
-            self.sidebar.setCellWidget(row_index, 1, delete_button)
-
         self.sidebar.styleRows()
         self.sidebar.selectRow(self.model.index)
 
@@ -1385,6 +1368,7 @@ class MainWindow(QMainWindow):
         if row != self.model.index:
             self.model.index = row
             self.data_changed()
+            self.sidebar.showCloseButton(row)
             self.model.history.append(f"data = datasets[{self.model.index}]")
 
     @Slot()